refactor: fixed API requests import
