+++ frontend/src/api/courses/index.ts
@@ -1,1 +1,3 @@
+export { default as useApplyCourse } from './applyCourse';
 export { default as useGetCourses } from './getCourses';
+export { default as useGetCourseInfo } from './getCourseInfo';
