+++ frontend/src/api/myCourses/getClientCourseInfo.ts
