+++ frontend/src/api/myCourses/index.ts
@@ -1,2 +1,5 @@
-export { default as useGetMyCourses } from './getMyCourses';
+export { default as useGetClientCourses } from './getClientCourses';
 export { default as useStartClientCourse } from './startClientCourse';
+export { default as useFinishClientCourse } from './finishClientCourse';
+export { default as useGetClientCourseInfo } from './getClientCourseInfo';
+export { default as usePassClientCourse } from './passClientCourse';
