+++ frontend/src/pages/LearningCourse/LearningCourseContainer.tsx
@@ -1,9 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router';
 
-import useGetClientCourseInfo from 'api/myCourses/getMyCourseInfo';
-import usePassClientCourse from 'api/myCourses/passClientCourse';
-import { useStartClientCourse } from 'api/myCourses';
+import { useStartClientCourse, useGetClientCourseInfo, usePassClientCourse } from 'api/myCourses';
 import { optimizeLink } from 'utils/helpers/videoPlayer/videoLink';
 import { getPreviewId } from 'utils/helpers/videoPlayer/getPreviewId';
 import { MATERIAL } from 'constants/materials';
