+++ frontend/src/pages/MyCourses/MyCoursesContainer.tsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react';
 
-import { useGetMyCourses } from 'api/myCourses';
+import { useGetClientCourses } from 'api/myCourses';
 import { WINDOW_SIZE } from 'constants/windowWidth';
 import { getWindowWidth } from 'utils/helpers/getWindowWidth';
 
 import MyCoursesList from './MyCoursesList';
 
 const MyCoursesContainer: React.FC = () => {
-  const { data: clientCourses, isLoading } = useGetMyCourses();
+  const { data: clientCourses, isLoading } = useGetClientCourses();
   const disableLinkWidth =
     window.innerWidth < WINDOW_SIZE.sm.width ? WINDOW_SIZE.xs.name : WINDOW_SIZE.sm.name;
   const disableLink = (event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
